AI_Database_chatbot_gemini/requirements.txt
psycopg2-binary
google-genai
vllm
aiohttp
bitsandbytes>=0.46.1